Telegrame_Bot-Lucifer--main/requirements.txt
python-telegram-bot
python-dotenv
requests
groq
cloudscraper
beautifulsoup4
edge-tts
qrcode[pil]
httpx
duckduckgo-search
google-genai
yt-dlp
Pillow
google-api-core
